node: remove modular classes
