+++ lib/index.ts
@@ -1,4 +1,4 @@
-import * as API from './Textile/API'
+import * as API from './Textile/Node'
 import Events, { TextileEvents } from './Textile/events'
 import Textile, { BackgroundTask } from './Textile'
 import * as util from './util'
